requirements.txt
Django==4.0.3
django-ckeditor==6.2.0
Pillow==9.0.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
django-widget-tweaks==1.4.12
requests==2.28.1